set default volume to f64

+++ src-tauri/migrations/2026-03-06-205856-0000_init/up.sql
@@ -108,7 +108,7 @@ CREATE TABLE
     settings (
         id INT NOT NULL PRIMARY KEY CHECK (id = 1),
         minimize_on_copy BOOLEAN NOT NULL DEFAULT 0 CHECK (minimize_on_copy IN (0, 1)),
-        default_volume FLOAT CHECK (
+        default_volume DOUBLE CHECK (
             default_volume >= 0.0
             AND default_volume <= 1.0
         )
